Check that cosign.pub matches private key
This avoids images which can't be rebased into due to `invalid
signature` errors because cosign.pub doesn't match the private key
actually used for signing. The error is caught early in the build
process as there's no point creating an image if cosign.pub is wrong.

Note: This means that cosign is installed for all builds instead of
conditionally for release builds.